refactor(endpoint): extract form, list, item from settingpage
