+++ src/components/admin/SettingPage/index.js
@@ -1,0 +1,7 @@
+import EndpointForm from './EndpointForm';
+import EndpointList from './EndpointList';
+
+export {
+    EndpointForm,
+    EndpointList
+}
